pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "smart-home"
version = "0.1.0"
description = "Agentic Smart Home client"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "requests>=2.31",
    "python-dotenv>=1.0",
    "vosk>=0.3.45",
    "sounddevice>=0.5.2",
    "pyttsx3>=2.99",
    "simpleaudio>=1.0.4",
    "openai>=2.7.1",
    "openwakeword>=0.6.0",
]
